Add image, description, ingredients, instructions
Went in to add the bulk of the recipe page. Added image into img folder and took it straight to the website

Then added a description using h2 for the header

Used h2 for both the ingredients and menu page while using unordered and ordered list properly. Also went ahead and added emphasis to the steps as they came with descriptions of their own.