Create 3 separate pages for website
Went ahead and put all the recipes on separate pages with main page having links to each

Also changed the links on each of the pages to the images. Had to go back out of the folder and go to image folder.

Also changed links to unordered list on main page